+++ run.go
@@ -7,9 +7,13 @@ import (
 
 // State tracks the current state of application execution.
 type State struct {
+	stateLock sync.RWMutex
+
 	machine      *machine
-	errors       <-chan error
 	shutdownOnce sync.Once
+
+	errors     <-chan error
+	errorsSeen []error
 }
 
 // Run builds a machine to invoke the processes registered to the given container. This
@@ -33,12 +37,25 @@ func Run(ctx context.Context, container *Container, configs ...MachineConfigFunc
 // returns a boolean flag indicating a clean exit.
 func (s *State) Wait(ctx context.Context) bool {
 	ok := true
-	for range s.errors {
+	for err := range s.errors {
 		ok = false
 		s.Shutdown(ctx)
+
+		s.stateLock.Lock()
+		s.errorsSeen = append(s.errorsSeen, err)
+		s.stateLock.Unlock()
 	}
 
 	return ok
+}
+
+// Errors returns a slice of errors encountered while running the processes. The Errors method can
+// be used once the Wait method returns to find the errors returned by the processes.
+func (s *State) Errors() []error {
+	s.stateLock.RLock()
+	defer s.stateLock.RUnlock()
+
+	return s.errorsSeen
 }
 
 // Shutdown signals all running processes to exit.