+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `State.Errors`.
+
 ## [v2.0.1] - 2022-10-10
 
 ### Added